update type and comment
