+++ src/components/NewDatePicker/NewDatePicker.tsx
@@ -12,10 +12,28 @@ import {
 } from "@floating-ui/react";
 
 export type NewDatePickerProps = {
+  /**
+   * 日付
+   * @default dayjs()
+   */
   date: Dayjs;
+  /**
+   * カレンダーの左に表示するアクション
+   */
   actions?: Action[];
+  /**
+   * 指定したい format
+   * @default YYYY-MM-DD
+   */
   format?: string;
+  /**
+   * エラーメッセージのテキスト
+   */
   errorText?: string;
+  /**
+   * 入力を無効にする
+   * @default false
+   */
   disabled?: boolean;
   /**
    * 選択可能なカレンダーの領域を制限する
@@ -23,6 +41,9 @@ export type NewDatePickerProps = {
    * @default () => false
    */
   isOutsideRange?: (date: Dayjs) => boolean;
+  /**
+   * 日付が変更されたときに呼ばれる関数
+   */
   onDateChange: (date: Dayjs) => void;
 };
