+++ src/components/NewDateRangePicker/NewDateRangePicker.tsx
@@ -18,10 +18,28 @@ import {
 import { DateRange } from "../Calendar/CalendarRange/types";
 
 export type NewDateRangePickerProps = {
+  /**
+   * 開始日
+   * @default dayjs()
+   */
   startDate: Dayjs;
+  /**
+   * 終了日
+   * @default dayjs()
+   */
   endDate: Dayjs;
+  /**
+   * エラーメッセージのテキスト
+   */
   errorText?: string;
+  /**
+   * カレンダーの左に表示するアクション
+   */
   actions?: Action[];
+  /**
+   * 入力を無効にする
+   * @default false
+   */
   disabled?: boolean;
   /**
    * 選択可能なカレンダーの領域を制限する
@@ -29,6 +47,9 @@ export type NewDateRangePickerProps = {
    * @default () => false
    */
   isOutsideRange?: (date: Dayjs) => boolean;
+  /**
+   * 日付が変更されたときに呼ばれる関数
+   */
   onDatesChange: (date: DateRange) => void;
 };
